Update README.md
Specifically, the word "heirarchical" should be written as "hierarchical", the phrase "mnemonics (typically 12, 18, or 24 word secret phrases)" should be written as "mnemonics (typically 12, 18, or 24-word secret phrases)", and the phrase "notice dashes versus underscores in the names of the utility and package" should be written as "note the difference between dashes and underscores in the names of the utility and package". Apart from that, the statement is correct.